xfstests: scoutfs support
- Make changes needed for two-device support in scoutfs
- Fix xfstests for new needed -f mkfs option
- Add scoutfs per-fs mkfs options

ScoutFS requires unique addresses for each quorum member slot for each
file system.  We add _TEST_, _SCRATCH_, and _DEV_ mkfs options for each
kind of fs that tests are going to try and make.  We also remove the
unconditional -f so that the config can put it in MKFS_OPTIONS.

Signed-off-by: Andy Grover <agrover@versity.com>
Signed-off-by: Zach Brown <zab@versity.com>
Signed-off-by: Auke Kok <auke.kok@versity.com>