Add scoutfs per-fs mkfs options
ScoutFS requires unique addresses for each quorum member slot for each
file system.  We add _TEST_, _SCRATCH_, and _DEV_ mkfs options for each
kind of fs that tests are going to try and make.  We also remove the
unconditional -f so that the config can put it in MKFS_OPTIONS.

Signed-off-by: Zach Brown <zab@versity.com>